fix redirect URL detection
