+++ internal/outpost/proxyv2/application/oauth_state.go
@@ -12,6 +12,7 @@ import (
 )
 
 type OAuthState struct {
+	Issuer    string `json:"iss" mapstructure:"iss"`
 	SessionID string `json:"sid" mapstructure:"sid"`
 	State     string `json:"state" mapstructure:"state"`
 	Redirect  string `json:"redirect" mapstructure:"redirect"`
@@ -20,26 +21,34 @@ type OAuthState struct {
 func (oas *OAuthState) GetExpirationTime() (*jwt.NumericDate, error) { return nil, nil }
 func (oas *OAuthState) GetIssuedAt() (*jwt.NumericDate, error)       { return nil, nil }
 func (oas *OAuthState) GetNotBefore() (*jwt.NumericDate, error)      { return nil, nil }
-func (oas *OAuthState) GetIssuer() (string, error)                   { return "goauthentik.io/outpost", nil }
+func (oas *OAuthState) GetIssuer() (string, error)                   { return oas.Issuer, nil }
 func (oas *OAuthState) GetSubject() (string, error)                  { return oas.State, nil }
 func (oas *OAuthState) GetAudience() (jwt.ClaimStrings, error)       { return nil, nil }
 
 var base32RawStdEncoding = base32.StdEncoding.WithPadding(base32.NoPadding)
 
-func (a *Application) createState(r *http.Request) (string, error) {
+func (a *Application) createState(r *http.Request, fwd string) (string, error) {
 	s, _ := a.sessions.Get(r, a.SessionName())
 	if s.ID == "" {
 		// Ensure session has an ID
 		s.ID = base32RawStdEncoding.EncodeToString(securecookie.GenerateRandomKey(32))
 	}
 	st := &OAuthState{
+		Issuer:    fmt.Sprintf("goauthentik.io/outpost/%s", a.proxyConfig.GetClientId()),
 		State:     base64.RawURLEncoding.EncodeToString(securecookie.GenerateRandomKey(32)),
 		SessionID: s.ID,
+		Redirect:  fwd,
 	}
-	rd, ok := a.checkRedirectParam(r)
-	if ok {
-		a.log.WithField("rd", rd).Trace("Setting redirect")
-		st.Redirect = rd
+	if fwd == "" {
+		// This should only really be hit for nginx forward_auth
+		// as for that the auth start redirect URL is generated by the
+		// reverse proxy, and as such we won't have a request we just
+		// denied to reference for final URL
+		rd, ok := a.checkRedirectParam(r)
+		if ok {
+			a.log.WithField("rd", rd).Trace("Setting redirect")
+			st.Redirect = rd
+		}
 	}
 	token := jwt.NewWithClaims(jwt.SigningMethodHS256, st)
 	tokenString, err := token.SignedString([]byte(a.proxyConfig.GetCookieSecret()))
@@ -60,6 +69,15 @@ func (a *Application) stateFromRequest(r *http.Request) *OAuthState {
 	})
 	if err != nil {
 		a.log.WithError(err).Warning("failed to parse state jwt")
+		return nil
+	}
+	iss, err := token.Claims.GetIssuer()
+	if err != nil {
+		a.log.WithError(err).Warning("state jwt without issuer")
+		return nil
+	}
+	if iss != fmt.Sprintf("goauthentik.io/outpost/%s", a.proxyConfig.GetClientId()) {
+		a.log.WithField("issuer", iss).Warning("invalid state jwt issuer")
 		return nil
 	}
 	claims := &OAuthState{}
